+++ pkg/clientset/typed/apis/v1alpha1/apis_client.go
@@ -30,6 +30,7 @@ type ApisV1alpha1Interface interface {
 	RESTClient() rest.Interface
 	AIGatewaysGetter
 	DataPlaneMetricsExtensionsGetter
+	KongPluginInstallationsGetter
 }
 
 // ApisV1alpha1Client is used to interact with features provided by the apis group.
@@ -43,6 +44,10 @@ func (c *ApisV1alpha1Client) AIGateways(namespace string) AIGatewayInterface {
 
 func (c *ApisV1alpha1Client) DataPlaneMetricsExtensions(namespace string) DataPlaneMetricsExtensionInterface {
 	return newDataPlaneMetricsExtensions(c, namespace)
+}
+
+func (c *ApisV1alpha1Client) KongPluginInstallations(namespace string) KongPluginInstallationInterface {
+	return newKongPluginInstallations(c, namespace)
 }
 
 // NewForConfig creates a new ApisV1alpha1Client for the given config.