+++ pkg/clientset/typed/apis/v1alpha1/generated_expansion.go
@@ -21,3 +21,5 @@ package v1alpha1
 type AIGatewayExpansion interface{}
 
 type DataPlaneMetricsExtensionExpansion interface{}
+
+type KongPluginInstallationExpansion interface{}